reorganize the queries endpoints to have some submodules and improve readability, add a pipeline query endpoint
