+++ api/src/routes/queries/cone_search.rs
@@ -1,10 +1,9 @@
 /// Endpoints for executing analytical queries.
 use crate::catalogs::catalog_exists;
-use crate::filters::{parse_filter, parse_optional_filter};
+use crate::filters::parse_optional_filter;
 use crate::models::response;
 
 use actix_web::{HttpResponse, post, web};
-use futures::StreamExt;
 use futures::TryStreamExt;
 use mongodb::{Database, bson::doc};
 use std::collections::HashMap;
@@ -12,159 +11,6 @@ use std::fmt;
 use utoipa::openapi::RefOr;
 use utoipa::openapi::schema::{ObjectBuilder, Schema};
 use utoipa::{PartialSchema, ToSchema};
-
-#[derive(serde::Deserialize, Clone, ToSchema)]
-struct CountQuery {
-    catalog_name: String,
-    filter: serde_json::Value,
-}
-
-/// Run a count query
-#[utoipa::path(
-    post,
-    path = "/queries/count",
-    request_body = CountQuery,
-    responses(
-        (status = 200, description = "Count of documents in the catalog", body = serde_json::Value),
-        (status = 404, description = "Catalog does not exist"),
-        (status = 500, description = "Internal server error")
-    ),
-    tags=["Queries"]
-)]
-#[post("/queries/count")]
-pub async fn post_count_query(
-    db: web::Data<Database>,
-    web::Json(query): web::Json<CountQuery>,
-) -> HttpResponse {
-    let catalog_name = query.catalog_name.trim();
-    if !catalog_exists(&db, &catalog_name).await {
-        return response::not_found(&format!("Catalog {} does not exist", catalog_name));
-    }
-    let collection_name = catalog_name.to_string();
-    // Get the collection
-    let collection = db.collection::<mongodb::bson::Document>(&collection_name);
-    // Count documents with optional filter
-    let filter = match parse_filter(&query.filter) {
-        Ok(f) => f,
-        Err(e) => return response::bad_request(&format!("Invalid filter: {:?}", e)),
-    };
-    let count = match collection.count_documents(filter).await {
-        Ok(c) => c,
-        Err(e) => {
-            return response::internal_error(&format!("Error counting documents: {:?}", e));
-        }
-    };
-    // Return the count
-    response::ok("success", serde_json::to_value(count).unwrap())
-}
-
-#[derive(serde::Deserialize, Clone, ToSchema)]
-struct EstimatedCountQuery {
-    catalog_name: String,
-}
-
-/// Run an estimated count query
-#[utoipa::path(
-    post,
-    path = "/queries/estimated-count",
-    request_body = EstimatedCountQuery,
-    responses(
-        (status = 200, description = "Approximately count documents in the catalog", body = serde_json::Value),
-        (status = 404, description = "Catalog does not exist"),
-        (status = 500, description = "Internal server error")
-    ),
-    tags=["Queries"]
-)]
-#[post("/queries/estimated-count")]
-pub async fn post_estimated_count_query(
-    db: web::Data<Database>,
-    web::Json(query): web::Json<EstimatedCountQuery>,
-) -> HttpResponse {
-    let catalog_name = query.catalog_name.trim();
-    if !catalog_exists(&db, &catalog_name).await {
-        return response::not_found(&format!("Catalog {} does not exist", catalog_name));
-    }
-    let collection_name = catalog_name.to_string();
-    // Get the collection
-    let collection = db.collection::<mongodb::bson::Document>(&collection_name);
-    let count = match collection.estimated_document_count().await {
-        Ok(c) => c,
-        Err(e) => {
-            return response::internal_error(&format!("Error counting documents: {:?}", e));
-        }
-    };
-    // Return the count
-    response::ok("success", serde_json::to_value(count).unwrap())
-}
-
-#[derive(serde::Deserialize, serde::Serialize, Clone, ToSchema)]
-struct FindQuery {
-    catalog_name: String,
-    filter: serde_json::Value,
-    projection: Option<serde_json::Value>,
-    limit: Option<i64>,
-    skip: Option<u64>,
-    sort: Option<serde_json::Value>,
-    max_time_ms: Option<u64>,
-}
-impl FindQuery {
-    /// Convert to MongoDB Find options
-    fn to_find_options(&self) -> mongodb::options::FindOptions {
-        let mut options = mongodb::options::FindOptions::default();
-        if let Some(projection) = &self.projection {
-            options.projection = Some(mongodb::bson::to_document(projection).unwrap());
-        }
-        if let Some(limit) = self.limit {
-            options.limit = Some(limit);
-        }
-        if let Some(skip) = self.skip {
-            options.skip = Some(skip);
-        }
-        if let Some(sort) = &self.sort {
-            options.sort = Some(mongodb::bson::to_document(sort).unwrap());
-        }
-        if let Some(max_time_ms) = self.max_time_ms {
-            options.max_time = Some(std::time::Duration::from_millis(max_time_ms));
-        }
-        options
-    }
-}
-
-/// Perform a find query on a catalog
-#[utoipa::path(
-    post,
-    path = "/queries/find",
-    request_body = FindQuery,
-    responses(
-        (status = 200, description = "Documents found in the catalog", body = serde_json::Value),
-        (status = 400, description = "Bad request"),
-        (status = 500, description = "Internal server error")
-    ),
-    tags=["Queries"]
-)]
-#[post("/queries/find")]
-pub async fn post_find_query(db: web::Data<Database>, body: web::Json<FindQuery>) -> HttpResponse {
-    let catalog_name = body.catalog_name.trim();
-    if !catalog_exists(&db, &catalog_name).await {
-        return response::not_found(&format!("Catalog {} does not exist", catalog_name));
-    }
-    let collection_name = catalog_name.to_string();
-    // Get the collection
-    let collection = db.collection::<mongodb::bson::Document>(&collection_name);
-    // Find documents with the provided filter
-    let filter = match parse_filter(&body.filter) {
-        Ok(filter) => filter,
-        Err(e) => return response::bad_request(&format!("Invalid filter: {:?}", e)),
-    };
-    let find_options = body.to_find_options();
-    match collection.find(filter).with_options(find_options).await {
-        Ok(cursor) => {
-            let docs: Vec<_> = cursor.map(|doc| doc.unwrap()).collect::<Vec<_>>().await;
-            response::ok("success", serde_json::to_value(docs).unwrap())
-        }
-        Err(e) => response::internal_error(&format!("Error finding documents: {:?}", e)),
-    }
-}
 
 #[derive(serde::Deserialize, Clone)]
 pub enum Unit {
@@ -248,7 +94,7 @@ impl ConeSearchQuery {
     }
 }
 
-/// Perform a cone search query on a catalog
+/// Run a cone search query on a catalog
 #[utoipa::path(
     post,
     path = "/queries/cone-search",
