+++ api/src/routes/queries/find.rs
@@ -1,0 +1,78 @@
+/// Endpoints for executing analytical queries.
+use crate::catalogs::catalog_exists;
+use crate::filters::parse_filter;
+use crate::models::response;
+
+use actix_web::{HttpResponse, post, web};
+use futures::StreamExt;
+use mongodb::{Database, bson::doc};
+use utoipa::ToSchema;
+
+#[derive(serde::Deserialize, serde::Serialize, Clone, ToSchema)]
+struct FindQuery {
+    catalog_name: String,
+    filter: serde_json::Value,
+    projection: Option<serde_json::Value>,
+    limit: Option<i64>,
+    skip: Option<u64>,
+    sort: Option<serde_json::Value>,
+    max_time_ms: Option<u64>,
+}
+impl FindQuery {
+    /// Convert to MongoDB Find options
+    fn to_find_options(&self) -> mongodb::options::FindOptions {
+        let mut options = mongodb::options::FindOptions::default();
+        if let Some(projection) = &self.projection {
+            options.projection = Some(mongodb::bson::to_document(projection).unwrap());
+        }
+        if let Some(limit) = self.limit {
+            options.limit = Some(limit);
+        }
+        if let Some(skip) = self.skip {
+            options.skip = Some(skip);
+        }
+        if let Some(sort) = &self.sort {
+            options.sort = Some(mongodb::bson::to_document(sort).unwrap());
+        }
+        if let Some(max_time_ms) = self.max_time_ms {
+            options.max_time = Some(std::time::Duration::from_millis(max_time_ms));
+        }
+        options
+    }
+}
+
+/// Run a find query on a catalog
+#[utoipa::path(
+    post,
+    path = "/queries/find",
+    request_body = FindQuery,
+    responses(
+        (status = 200, description = "Documents found in the catalog", body = serde_json::Value),
+        (status = 400, description = "Bad request"),
+        (status = 500, description = "Internal server error")
+    ),
+    tags=["Queries"]
+)]
+#[post("/queries/find")]
+pub async fn post_find_query(db: web::Data<Database>, body: web::Json<FindQuery>) -> HttpResponse {
+    let catalog_name = body.catalog_name.trim();
+    if !catalog_exists(&db, &catalog_name).await {
+        return response::not_found(&format!("Catalog {} does not exist", catalog_name));
+    }
+    let collection_name = catalog_name.to_string();
+    // Get the collection
+    let collection = db.collection::<mongodb::bson::Document>(&collection_name);
+    // Find documents with the provided filter
+    let filter = match parse_filter(&body.filter) {
+        Ok(filter) => filter,
+        Err(e) => return response::bad_request(&format!("Invalid filter: {:?}", e)),
+    };
+    let find_options = body.to_find_options();
+    match collection.find(filter).with_options(find_options).await {
+        Ok(cursor) => {
+            let docs: Vec<_> = cursor.map(|doc| doc.unwrap()).collect::<Vec<_>>().await;
+            response::ok("success", serde_json::to_value(docs).unwrap())
+        }
+        Err(e) => response::internal_error(&format!("Error finding documents: {:?}", e)),
+    }
+}
